Set psql settings to defaults, and don't force a user
A user might have a psqlrc file which sets non-default values, which
can cause problems for trying to parse the psql response. These
flags are suitable for parsing the results in a bash script

Avoid forcing a particular user. If a particular user is needed
this can be set with PGUSER